requirements.txt
bleak>=0.21.0
mindspore>=2.3.0
numpy>=1.24.0
rich>=13.0.0
pandas>=2.0.0
llama-cpp-python>=0.2.0
